requirements.txt
arrow==0.8.0
blessed==1.14.1
Django==1.10
django-appconf==1.0.2
django-bootstrap3==7.0.1
django-picklefield==0.3.2
django-q==0.7.18
django-user-accounts==1.3.1
future==0.15.2
hashids==1.1.0
psycopg2==2.6.2
python-dateutil==2.5.3
pytz==2016.6.1
requests==2.11.1
six==1.10.0
wcwidth==0.1.7
whitenoise==3.2.2
